% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_presentation.R
\name{add_presentation}
\alias{add_presentation}
\title{Create or add to a \code{seminaR} object}
\usage{
add_presentation(object, presenter = NA, presenter_name = NA,
  presenter_midname = NA, presenter_surname = NA,
  presenter_affiliation = NA, presenter_email = NA,
  presenter_link = NA, country = NA, city = NA, lon = NA,
  lat = NA, event = NA, title = NA, abstract = NA, type = NA,
  link = NA, institution = NA, department = NA, school = NA,
  venue = NA, room = NA, year = NA, month = NA, day = NA,
  start_hour = NA, end_hour = start_hour, start_min = 0L,
  end_min = start_min, start_sec = 0L, end_sec = 0L, tz = "UTC")
}
\arguments{
\item{object}{an object an object of class \code{"seminaR"}.}

\item{presenter}{either \code{NA} (default) or an object of class \code{\link{seminaR_presenter}}. In the latter case, all input to the \code{presenter_*} arguments below is ignored and populated according to the supplied object.}

\item{presenter_name}{name of the presenter of the talk/seminar; character string or \code{NA} (default).}

\item{presenter_midname}{middle name of the presenter of the talk/seminar; character string or \code{NA} (default).}

\item{presenter_surname}{surname of the presenter of the talk/seminar; character string or \code{NA} (default).}

\item{presenter_affiliation}{affiliation of the presenter of the talk/seminar; character string or \code{NA} (default).}

\item{presenter_email}{email of the presenter of the talk/seminar; character string or \code{NA} (default).}

\item{presenter_link}{link to the webpage of the presenter of the talk/seminar; character string or \code{NA} (default).}

\item{country}{country of the talk/seminar; character string or \code{NA} (default).}

\item{city}{city of the talk/seminar; character string or \code{NA} (default).}

\item{lon}{longitude of the venue of the talk/seminar; numeric or \code{NA} (default).}

\item{lat}{latitude of the venue of the talk/seminar; numeric or \code{NA} (default).}

\item{event}{name of the event at which the talk/seminar is/was given; character string or \code{NA} (default).}

\item{title}{title of the talk/seminar; character string or \code{NA} (default).}

\item{abstract}{abstract of the talk/seminar; character string or \code{NA} (default).}

\item{type}{the type of the talk. Available options are \code{NA} (default), \code{"seminar"}, \code{"lecture"}, \code{"presentation"}, \code{"talk"}, \code{"poster"}.}

\item{link}{link to the event or seminar/talk page; character string or \code{NA} (default).}

\item{institution}{institution at which the event or seminar/talk page took/will take space; character string or \code{NA} (default).}

\item{department}{department at which the event or seminar/talk page took/will take space; character string or \code{NA} (default).}

\item{school}{school at which the event or seminar/talk page took/will take space; character string or \code{NA} (default).}

\item{venue}{venue at which the event or seminar/talk page took/will take space; character string or \code{NA} (default).}

\item{room}{room at which the event or seminar/talk page took/will take space; character string or \code{NA} (default).}

\item{year}{year of the talk/seminar; numeric, e.g. \code{2019}, or \code{NA} (default).}

\item{month}{month of the talk/seminar; numeric (1-12) or \code{NA} (default).}

\item{day}{day of the talk/seminar; numeric (1-31) or \code{NA} (default).}

\item{start_hour}{start hour of the talk/seminar; numeric (1-24) or \code{NA} (default).}

\item{end_hour}{end hour of the talk/seminar; numeric (1-24) or \code{NA} (default).}

\item{start_min}{start minute of the talk/seminar; numeric (0-60) or \code{NA} (default).}

\item{end_min}{end minute of the talk/seminar; numeric (0-60) or \code{NA} (default).}

\item{start_sec}{start second of the talk/seminar; numeric (0-60) or \code{NA} (default).}

\item{end_sec}{end second of the talk/seminar; numeric (0-60) or \code{NA} (default).}

\item{tz}{timezone. Default is \code{"UTC"}. See \code{\link{DateTimeClasses}} for details.}
}
\value{
A structured \code{\link{data.frame}} that also inherits from class \code{seminaR}, including the supplied talk/seminar details.
}
\description{
Create or add to a \code{seminaR} object
}
\details{
If \code{object} is not specified then \code{add_talk} will create an \code{seminaR} objects based on the supplied inputs, otherwise it will add the details of the new talk/seminar on \code{object}.
}
