% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.seminaR}
\alias{plot.seminaR}
\title{Interactive seminar maps}
\usage{
\method{plot}{seminaR}(x, group = "year", title = NA,
  title_position = "bottomleft", provider = "OpenStreetMap.Mapnik",
  interval = TRUE, date_format = "dmy", shorten_URLs = FALSE,
  service = "Is.gd", ...)
}
\arguments{
\item{x}{an object of class \code{"seminaR"}. See \code{\link{add_presentation}}.}

\item{group}{according to what should the seminars be selected on the map? Available options are \code{"year"} (default), \code{"month"}, \code{"presenter"}, \code{"event"}, \code{"country"}.}

\item{title}{character string for the title of the map. Default is \code{NA}, which produces no title.}

\item{title_position}{the position of the title on the map, if \code{title} is not \code{NA}. Available options are \code{"bottomleft"} (default), \code{"bottomright"}, \code{"topleft"}, \code{"topright"}.}

\item{provider}{the provider of tiles for the base map. See \code{\link[leaflet]{addProviderTiles}}. Default is \code{"OpenStreetMap.Mapnik"}.}

\item{interval}{Should the start and end times in each popup be displayed as an interval (\code{TRUE}; default) or in two separate lines (\code{FALSE})?}

\item{date_format}{In what format should the dates be displayed? Available options are \code{"dmy"}, \code{"mdy"}, \code{"ydm"}, \code{"ymd"}, where \code{"y"} stands for year, \code{"m"} stands for month, and \code{"d"} stands for day.}

\item{shorten_URLs}{Should the URL links in \code{"object$link"} be shortened? Default is \code{FALSE}.}

\item{service}{service to use for shortening URLs. Current options are \code{"Is.gd"} (default) and \code{"V.gd"}. See \code{\link{shorten_URLs}}.}

\item{...}{Arguments to be passed to other methods. Currently unused.}
}
\description{
Interactive seminar maps
}
