% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presenter.R
\name{presenter}
\alias{presenter}
\title{Create a \code{seminaR_presenter} object with presenter details}
\usage{
presenter(name = NA, midname = NA, surname = NA, affiliation = NA,
  link = NA, email = NA)
}
\arguments{
\item{name}{name of the presenter; character string or \code{NA} (default).}

\item{midname}{middle name of the presenter; character string or \code{NA} (default).}

\item{surname}{surname of the presenter; character string or \code{NA} (default).}

\item{affiliation}{affiliation of the presenter; character string or \code{NA} (default).}

\item{link}{link to the webpage of the presenter; character string or \code{NA} (default).}

\item{email}{email of the presenter; character string or \code{NA} (default).}
}
\description{
Create a \code{seminaR_presenter} object with presenter details
}
