% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seminaR-package.R
\docType{package}
\name{seminaR}
\alias{seminaR}
\alias{seminaR-package}
\title{seminaR: Methods and classes for maintaining a database of talks and seminars}
\description{
seminaR: Methods and classes for maintaining a database of talks and seminars
}
\details{

}
