% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semnar-package.R
\docType{package}
\name{semnar-package}
\alias{semnar-package}
\title{semnar: Methods and classes for constructing, maintaining and interacting with a database of presentations}
\description{
semnar: Methods and classes for constructing, maintaining and interacting with a database of presentations
}
\seealso{
\code{\link{add_presentation}} \code{\link{presenter}} \code{\link{plot.semnar}}
}
