% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semnar-package.R
\docType{package}
\name{semnar}
\alias{semnar}
\alias{semnar-package}
\title{semnar: Methods and classes for maintaining a database of talks and seminars}
\description{
semnar: Methods and classes for maintaining a database of talks and seminars
}
