% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seminaR-package.R, R/shorten_URLs.R
\name{shorten_URLs}
\alias{shorten_URLs}
\alias{shorten_URLs.seminaR}
\title{Shorten the URLs of any links in a \code{"seminaR"} object}
\usage{
shorten_URLs(object, service = "V.gd")

\method{shorten_URLs}{seminaR}(object, service = "Is.gd")
}
\arguments{
\item{object}{an object of class \code{"seminaR"}. See \code{\link{add_presentation}}.}

\item{service}{service to use for shortening URLs. Current options are \code{"V.gd"} (default) and \code{"Is.gd"}.}
}
\description{
Shorten the URLs of any links in a \code{"seminaR"} object
}
